object-storage: cleanup err handling in Glusterfs
Clean up the error handling in this module. This is a back-port of
work done as part of upstream refactoring work to remove the
dependency on the Swift differences we carry, see commit
6194691528d487417288e4dd0fa377fa06e85a04.

This reduces a number of errors seen at customer sites that just added
noise to the log files since the errors are recoverable situations.

See BZ 904370 (https://bugzilla.redhat.com/show_bug.cgi?id=904370).

Change-Id: Ic759fcab47c24e890a8655f0cdd79ffe14064330
Signed-off-by: Peter Portante <peter.portante@redhat.com>