object-storage: turn off act and cont updates
For gluster, we don't need to update an account or a container when a PUT
operation occurs for a container or object, respectively.

Since the account is just the gluster volume (one-to-one mapping of account to
gluster volume mount point), the PUT of a container just creates a directory
in the root of the volume, so it has already "put" it into the account.

Since the object is just a file or directory placed into a container which, in
turn, is just a directory, the act of creating it has already "put" it into
the container.

Removing these two operations saves a round trip on PUT operations for
containers and objects, which is a big savings for objects.

See BZ 903396 (https://bugzilla.redhat.com/show_bug.cgi?id=903396).

Change-Id: Ic9b37defef3611d9ef342a17cbfe8040df9cca4e
Signed-off-by: Peter Portante <peter.portante@redhat.com>