object-storage: add worker max_clients parameter
Adding the "max_clients" parameter for a worker allows us to handle
one client at a time for a worker. This is very useful for debugging
behaviors.

But it is also useful for working around a starvation issue when the
file system underneath becomes very slow. By setting max_clients to 1,
and bumping up the number of workers significantly, we can attempt to
mitigate that starvation effect.

See BZ 844805 (https://bugzilla.redhat.com/show_bug.cgi?id=844805).

Change-Id: I4d7db1e0c0eb05da73f19d5abd5f32e4cec11475
Signed-off-by: Peter Portante <peter.portante@redhat.com>