object-storage: update spec files to match RHS 2.0
Update the spec files to match what is shipped with RHS 2.0 and add a
trailer to the revision to mark these as special builds "pdq".

This was performed to be sure test RPMs handed to customers don't
conflict with the original product builds. See the following BZ
covering most of those customer issues: BZ 903396
(https://bugzilla.redhat.com/show_bug.cgi?id=903396).

Change-Id: I408945ad69398cb7ed3d8511e0d64b63d1a5d51e
Signed-off-by: Peter Portante <peter.portante@redhat.com>