object-storage: don't perform listing for HEAD req
For both the account and container HEAD listing APIs don't walk the
entire tree to update the return values, as they don't need to be
accurate (for now). By removing the tree walk and accompanying flood
of stat() system calls, the system does not become stressed as a
result.

This was a change the original Gluster/Swift integration created, this
is not part of the behavior of Swift (since each PUT/DELETE operation
on a container/object update a single row in either the account or
container databases).

Note this change does not use the accurate_size_in_listing behavior
parameter since that is intended for GET requests on
containers/accounts.

See BZ 903396 (https://bugzilla.redhat.com/show_bug.cgi?id=903396).

Change-Id: I702f58c2c0058cfa74101cf38ba4d24c5aaae3f7
Signed-off-by: Peter Portante <peter.portante@redhat.com>