object-storage: avoid fetching metadata for lists
Only fetch the metadata for objects in a container, or containers in an
account, when the REST API asks for them (using an XML or JSON content-type).

We save many thousands of stat() and xattr system calls as a result.

See BZ 903396 (https://bugzilla.redhat.com/show_bug.cgi?id=903396).

Change-Id: I053e5421f36beb5b947ca5add4e9a0c727dc2488
Signed-off-by: Peter Portante <peter.portante@redhat.com>