object-storage: work around for mkdir() race condition
This is a work around for a race condition in certain Gluster versions
where two or more threads issuing an mkdir() system call for the exact
same directory can end up in D-state, symptomatic of the created
directory having two different gfids.

The work around uses a file lock applied to the parent directory to
ensure that only one thread can issue an mkdir for a directory in the
parent.

BUG: 906966 (https://bugzilla.redhat.com/show_bug.cgi?id=906966)
Change-Id: Ifb9b210821828e2f28863f6756efbff9e6885b3b
Signed-off-by: Peter Portante <peter.portante@redhat.com>