object-storage: add config option mount per server
Add a configuration option to allow each object server to use a unique
mount point for its operations on the file system. This gives us a bit
of a boost for both bandwidth and IOps.

We create a lock file for each unique instance of an object-server
using a lock file, and when a server acquires that lock file, it can
use or create the corresponding numbered mount point.

Change-Id: I538241711d0f542971e467b663799332a5b8b09f
Signed-off-by: Peter Portante <peter.portante@redhat.com>